backend/requirements.txt
# Statistical computing
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4
numpy>=1.25.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
lifelines>=0.27.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Web and CORS
python-multipart==0.0.6

# Authentication and Database
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0

# Image processing for better matplotlib output
pillow>=10.0.0

# Additional scientific packages
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Security
bcrypt>=4.1.0
cryptography>=41.0.0

# Async support
asyncpg>=0.29.0
httpx>=0.25.0 